"4905e4ff40d43ced265dbcd1d5511b86c8cd8f0c - tools/gzbd: Do not store values to the GtkListStore that are never fetched
gtk_tree_model_get() is used to explicitly get values from a GtkListStore.
The only value we ever explicitly fetch from the GtkListStore is
DZ_ZONE_NUM.

For all columns, there is an attibute specified.
This attribute is used to implicitly fetch a value for the key store.
However, if a column has a cell data function, then the cell data
function is used instead of the attribute.
Specifying an attribute when there is a cell data function is simply
misleading, since it will never be used.

None of the cell data functions ever explicitly fetch a value other
than DZ_ZONE_NUM from the GtkListStore.

There is no need to continuously save these attributes, since
they are never fetched, implicitly or explicitly.

Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>
"