"feb5229b4ca16e315f20fbad62d62fcb0e294324 - lib: fix BLKGETZONESZ use compilation errors
Fix compilation errors with kernel versions preceding 4.20 that do not
define the BLKGETZONESZ ioctl command.
Introduce the helper function zbd_get_zone_sectors() to obtain a
device zone size in number of 512B sectors. This helper is conditionally
defined depending on the existence of the BLKGETZONESZ ioctl call. For
kernel versions not defining this ioctl, the sysfs chunk_sectors
attribute is used to obtain the device zone size.

This helper is used in zbd_do_get_info() to initialize the device
information zone_sectors and zone_size fields.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>
"