"10f651308ffe64bf4d4acaf74ec68e609d8d7f6d - lib: Fetch max open/active zones from sysfs
Fetch max open/active zones from sysfs and save it in the
device information data structure.

Remove the special handling of unknown.
The sysfs API for max open/active zones does not differentiate
between "no limit" and "unknown", therefore, neither can we.

If we are running against a kernel that does not export these
properties to sysfs, we cannot know the limit, therefore,
default these values to "no limit".

Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>
"